Re-organize tests to uses ReTest
